@@ -1,7 +1,7 @@
 import styled from 'styled-components'
 import Exit from '../img/exit.png'
 import Home from '../img/home.jpg'
-import {useHistory} from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
 const Containter = styled.div`
 	display: flex;
@@ -13,20 +13,20 @@ const Containter = styled.div`
 
 
 const Footer = ()=>{
-	const history = useHistory()
+	const history = useNavigate()
 
 	const logout = ()=>{
 		const decide = window.confirm('Tem certeza que deseja deslogar?')
 
 		if(decide){
 			localStorage.removeItem('token')
-			history.push('/')
+			history('/')
 		}
 	}
 
 
 	return<Containter>
-			<img src={Home} onClick={()=> history.push('/')} alt=''/>
+			<img src={Home} onClick={()=> history('/')} alt=''/>
 			<img src={Exit} onClick={logout} alt=''/>
 		  </Containter>
 }
