@@ -4,12 +4,12 @@ import {url} from '../../constants/urls'
 import {Container} from './styled'
 import Header from '../../components/Header'
 import Footer from '../../components/Footer'
-import {useHistory} from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
 
 //===========================Inicio do compoente funcional=========
 const Transfer = ()=>{
-	const history = useHistory()
+	const history = useNavigate()
 	const [form, setForm] = useState({
 		name:'',
 		cpf:'',
@@ -22,10 +22,10 @@ const Transfer = ()=>{
 		const token = localStorage.getItem('token')
 
 		if(token === null){
-			history.push('/')
+			history('/')
 		}
 
-	}, [history])
+	}, [])
 
 	const onChange = (e)=>{
 		const {name, value} = e.target
