@@ -1,1 +1,1 @@
-export const url = 'http://localhost:3003'
+export const url = 'https://labebank-frontend.herokuapp.com'
