@@ -1,4 +1,4 @@
-import {Switch, Route} from 'react-router-dom'
+import {Routes, Route} from 'react-router-dom'
 import Balance from '../pages/Balance/Balance'
 import Deposit from '../pages/Deposit/Deposit'
 import Transfer from '../pages/Transfer/Transfer'
@@ -7,37 +7,19 @@ import Statement from '../pages/Statement/Statement'
 import Home from '../pages/Home/Home'
 import Login from '../pages/Login/Login'
 import Signup from '../pages/Signup/Signup'
-import GlobalState from '../global/GlobalState'
 
 
-const Routes = ()=>{
-	return<Switch>
-			<GlobalState>
-				<Route exact path='/'>
-					<Home/>
-				</Route>
-				<Route exact path='/login'>
-					<Login/>
-				</Route>
-				<Route exact path='/signup'>
-					<Signup/>
-				</Route>
-				<Route exact path='/balance'>
-					<Balance/>
-				</Route>
-				<Route exact path='/deposit'>
-					<Deposit/>
-				</Route>
-				<Route exact path='/transfer'>
-					<Transfer/>
-				</Route>
-				<Route exact path='/pay'>
-					<Payments/>
-				</Route>
-				<Route exact path='/statement'>
-					<Statement/>
-				</Route>
-			</GlobalState>
-		  </Switch>
+
+const Router = ()=>{
+	return<Routes>
+			<Route exact path='/' element={<Home/>}/>
+			<Route exact path='/login' element={<Login/>}/>
+			<Route exact path='/signup' element={<Signup/>} />
+			<Route exact path='/balance' element={<Balance/>}/>
+			<Route exact path='/deposit' element={<Deposit/>}/>
+			<Route exact path='/transfer' element={<Transfer/>}/>
+			<Route exact path='/pay' element={<Payments/>}/>
+			<Route exact path='/statement' element={<Statement/>}/>
+		  </Routes>
 }
-export default Routes
+export default Router
